basic functionality of server and client
I added the base functions for creating the server and clients. Currently, the server starts, and clients can connect to it. They are prompted for a username, however, the server is failing to read that username. I also tried setting up the server's function to notify the other users that someone joined, but it is not wokring yet.